wego/map-services-repo/requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.84
botocore==1.34.84
CacheControl==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.13
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iso3166==2.1.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mapbox==0.18.1
MarkupSafe==2.1.5
msgpack==1.0.8
openapi-codec==1.3.2
packaging==24.0
polyline==2.0.2
pymongo==3.12.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
s3transfer==0.10.1
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
